Dokerize + make it callable multiple times

+++ src/vps_status_common/db_helper.py
@@ -14,12 +14,15 @@ class DbHelper:
 
         self.con = sqlite3.connect(sqlite_db_file)
         self.cur = self.con.cursor()
-        self.cur.execute("DROP TABLE IF EXISTS status")  # todo: update instead of drop ...
 
-        self.cur.execute(f"CREATE TABLE status ( {self.headers})")
+        status_columns = self.cur.execute("SELECT name FROM PRAGMA_TABLE_INFO('status');").fetchall()
+        if len(status_columns) != (len(fi) + 1):
+            self.cur.execute("DROP TABLE IF EXISTS status")  # todo: update instead of drop ...
+
+        self.cur.execute(f"CREATE TABLE IF NOT EXISTS status ( id integer primary key autoincrement, {self.headers})")
 
     def save_status(self, status: Status):
-        query = f"INSERT INTO status VALUES({self.placeholders})"
+        query = f"INSERT INTO status ({self.headers}) VALUES({self.placeholders})"
         self.cur.execute(query, astuple(status))
 
     def save_vps(self, vps: VPS):
